Debug mode and RenderInfo component
Start using context for toggling debug mode, but that way it was
impossible to detect whether a render was from switching debug mode
when RenderInfo is used inside a memoed component.

Instead I made use of the already present ID to register a handler that
updates each components state when a global `debugMode` is changed.
Increment amount of debug mode changes, and assign it to each
component's state to force a render. Components observe whether the
debug mode changed during render to avoid incrementing and keep the
count accurate with the parent component.

Set a root CSS custom property to track the max amount of renders of any
rendered component. Then divide each element's amount of renders by
this, and use it in the background opacity. This has the effect of
highlighting the most frequently rendered elements.