Move global settings to custom hook
And move some more things out of theme editor, mostly history buttons.

The settings hook is also a place to manage sending messages to the
frame and adding hotkeys.

I tried to write this in a way that you only had to specifiy the name
once in an object key, and then generate setter name and options key
name from this. This worked and though it was very clean, it's not great
for code completion. Also in case an effect or hotkey is used, getter
and setter are usually both needed anyway.

Various small fixes.