Movable panels v1
* Use a portal to one of the other areas to make the element render
there instead.
* Position is kind of wonky. Every render will append at the end when
you drop. Ordering therefore should be done using CSS order property,
otherwise it would be hard to make it render in the right order without
adding wrapper elements.
* Areas have an id, but for area content this is more challenging.
Perhaps using React.Children to access some of the element's properties
like id or className to generate an identifier that is likely to be
unique inside the area.
* Instead, it currently uses the index of the element. This works well
but makes it very hard to preserve UI state when code elements are added
or remove.