History using resumable state v1
In this commit history behavior is extracted from the useThemeEditor hook.

The goal is that multiple components should be able to participate
in the same linear history.

This works but only a few components were added for now.

Eventually it should be possible to give in application control
over which aspects should be included in history.

In order to do this a new function called `useResumableReducer` is added.
This is a drop in replacement for `useReducer`. It additionally accepts
an ID to assist components that might dis- and re-appear.

### What has to still change?

For now it causes an extra render because the "main" reducer
is implemented with `useReducer`.

I only did this because it was a manageable path forward,
given the shape of the code I already had.
(i.e. changing this was a lot of work on uncommitted code)

That said, these extra renders don't have much impact.
They're mostly annoying while debugging.

Still seems like `useSyncExternalStore` should be able to do the
same without it.