Fix extremely weird bug where background style not updated
I observed tearing here, however it's likely related to a bug in React 18.

When pressing "revert" on a variable, the title attribute would
reflect the latest state (the default value),
but not the background property.

Changing this to set the non-shorthand properties fixed that.
It does require setting potentially invalid values.
But that already happened anyway and they'll just be ignored.

I don't exclude that it's related to the new state management
that uses `useSyncExternalStore`, but it seems unlikely.
These properties are both set from `value` in the same scope,
there seems no opportunity for tearing here.