useInsertionEffect twice
Once for updating the style, which seems the obvious
use case.

The other I'm not 100% confident about, but it seems
to work and likely performs better.
It's similar to the CSS in JS case in that it also affects styling.
Technically it shouldn't affect appearance, because of  the use of order,
but the browser can't know that.