Move simple sizes state out of theme editor
This is the first step of moving top level app state
into functions which can be called any number of times.

Since one hook calls the other, and a component calls both,
one hook may be called any number of times on the same component.
It might add duplicate notifiers, but it's unlikely
to have any significant impact.