TypeScript first steps
I had long awaited a good use case to pick up TypeScript,
finally the "easy accessor" get function is exactly that.

With this in place, you can access the value of hooks that you
don't intend writing to in the simplest possible way.
Without TS, there was no auto completion for the get function.