Prevent PicoCSS plans to make world burn with 100% wide default buttons
All previous examples of frameworks I loaded had styles that were at most a bit hairy,
but generally the result of mashing them together with my editor styles was sensible.
100% width buttons... not so much.
Still I'm refusing to separate the editor styles now, as it usually provides a quite nice result giving consistency with the page it's styling.
It's also handy in getting to a set of essential required CSS.