Better source maps step 1: parse compiled CSS
The position of certain elements in the compiled files needs to be known
to lookup the original source position for something.

I naively started with a regex, planning to extract indexes from the matches.
However simply looping through the chars seems much more performant.
I preserved the regex version for comparison.

The end result should allow to use the source map for all kinds of entities.

The algorithm could be simplified.