Voice commands: first steps
Use the shady SpeechRecognition API that nevertheless (barely) works.

## Features
* Menu structure to reach commands by voice
* Commands with parameters
* Chain menu navigation and commands with parameters in a single sentence
* First commands: scrolling a (non root) element
* React component to display current menu options

## Issues
### No grammar :(
Though the specs mention the ability to add a grammar to avoid it sending any string,
this is actually a facade API that was never implemented.
In Chrome, it just keeps sending any string it recognizes.
Other browsers simply don't have the API.
This forces you to choose words that it doesn't get wrong.
For example, using "end" (for scroll to bottom) didn't work,
it was always captured as "and".

## Numbers randomly captured either as word or numerical string
Which way depends on context, but there's no clear relation.
Will fix by adding mapping (as it only seems to happen on numbers < 10).

## Saying a word more than twice in a row stops working
If the string is identical 2 times in a row, the recognition stops working for that string,
until another string is used first.
Probably a bug in Chrome's implementation. Maybe related to excluding
background voice, like music?

## Microphone auto-disables after silent period
This makes it hard to constantly have voice commands available.
It requires an interaction to start listening again.
For regular mic access this is solvable, but not sure if speech
recognition allows this too.
The best workaround so far was to ensure there was some sound to reset the timer.
This could be any sound, not just spoken.
Perhaps it can use speakers to produce this sound.
Or request the user to put on some instrumental music.