Remove code to inspect outside of frame
Initially the theme editor did not use a frame, and instead was a floating
div that inspected the same page it was on.

I preserved this for as long as it didn't require too much code, as it's nice
to be able to inspect and modify the theme editor itself.

However, once inspected elements got added to the history timeline,
this became too contrived to support.

In fact, just prior to this commit it was already not functioning anymore.