Include raw values (WIP), filter selectors (WIP)
This includes every distinct raw value as though it's a separate
custom property.
It's kind of a hack, but worth it, given the simplicity.
Without it, some overridden rules are shown anyway.

Such values can only be viewed.
Parsing is kept simple, and has a bug in some spacing cases,
where you'll see the variable again as a raw value.

Also includes the start for filtering selectors in the UI.
Filtering for this cases is not ideal, as it makes the UI
you select the filters in jump around.
Probably a darker overlay for filtered elements would work better.