WIP improve default values and references
This contains all kinds of changes, since the code
is largely being rewritten.

Some things are not touched yet because the new
CSS parser will replace (or make obsolete) a lot of old logic,
once it's ready to put in place.

While this commit introduces some new bugs, it fixes way more.

Current theme view is restored, but not fully functional.