Multiple quick fixes, this code will largely dissappear soon anyway
Some of this is not pretty and makes the existing code even worse.
But it seems worth it, each of these things fix some nasty bugs.

Also includes a small amount of code needed for drag and drop of values,
the rest is added in a later commit.

- remove some root scope special treatment
- exclude currentcolor in preview swatch (dirty but it was a real eye sore)
- attempt using max specific property (only makes sense in current weird inspection logic)
- attempt displaying preview for a var that is the complete set of arguments of a css function
(only works if there are no additional args, but it's better than nothing)