WIP Drag styles onto page directly
First steps to support drag and drop onto page elements.
The main challenge is how to choose between different options
the dropped value could go to.
There's a lot of different ways this could be handled.
For now it's not handled at all, and just picks the first option,
with no way yet to switch to the other ones.

It also shows a select list in the history of such an action,
but it doesn't do anything but shows the alternatives.