Experiment: dynamic tabs, making contenteditable "work"
I might use this component to create a "multi-palette" component.
So I tested if it works with dynamic React elements,
which seems to be the case.

While at it, I also experimented with contenteditable.
If you use the simplest and intuitive implementation,
it leads to a messed up cursor position after input.

I tried restoring the selection range, but that's incredibly hairy,
and couldn't get it to work yet.

Instead, it seems way easier to exclude the element from rendering,
relying on the fact that nothing else is changing this value.

After I wrote that using `memo`, I realized that it can
possibly be handled using an "uncontrolled" component.
Though I didn't try, perhaps it doesn't work on contenteditable.