Consider as <color> if any usage only accepts that
Previously, it only considered whether the matching usage
of the current inspected element could be a color.
Combined with the fact that types of parts of shorthands
cannot yet be determined, this resulted in a lot of color
variables not getting the color control.

This change dramatically reduces the occurrence of this,
though it's still possible that only shorthands are present.