Visualize SVG parts
This is unexpectedly useful. It makes it easy to see how some
types of SVG are built.
Also useful when there's CSS on SVG parts, though this is rare.
An example is the sun and sun beams on OpenProps page,
and the header link logo.

Usually works well but doesn't properly display some cases.