History locking, step 1
Allow locking a state so that it can be ignored while browsing history.
This allows for things like stepping through your changes with multiple
screen sizes.

For now, to keep it simple, it only applies the lock when in an older state.

Locks on the latest state don't work yet.