Avoid losing history or prompting user by storing last "alternate"
The user is still prompted, but only when about to lose the
stashed changes.
This makes splicing edits into history a seamless experience,
that only involves a single button click once the new changes were added,
which restores the history by replaying all actions.

There are still some poorly handled cases when locks
are used together with this.