Ensure fast initial rendering of potentially large list of variables
By determining a window to show around the active element.
At least on Chromium, the browser handles changes to the list on both
sides of the initial elements very well: the insertion is only noticeable if you look at the scrollbars.
I had no idea it was this simple.

Because of this change, you can now scroll fast through history without
causing stutter as multiple instances of this list get added and removed.