Improve hack that provides missing custom prop values to previews
Turns out it's INCREDIBLY expensive to set large amounts of custom properties in React.

In the halfmoon demo it resulted in several hundreds ms extra time to render.
While it did result in tens of thousands of properties being set, it still seems longer than it should take.
Luckily this is all just a temporary hack and eventually will not be needed at all when references are properly tracked.

Set them on the group element to do this the minimum amount of times,
even though it does expand the hacky situation of applying the styles to the editor itself.
Also exclude root level properties as they are already accounted for by said hacky situation.