Improve behavior of locking reducer based state, improve locks+stash combo
Previously, it would continue using the state at the locked entry,
but only kept the actions of that entry, and not others before it.

As a result, you'd temporarily see the results of actions that are gone.

To fix this, it now creates entries for all actions leading up to the locked entry.

When a stash is created, it now filters out entries that were also recreated to preserve locks.