WIP  `calc` first steps, evaluate expression under various conditions
Parse values that are a `calc` expression and evaluate the result.
For now, it doesn't create a visual representation of the inner operations,
just a text field to keep things simple.

For now only size controls use this to make it simpler to introduce,
and it's quite rare to see calc used outside of this.

Working well:
- all math operators available in the spec
- math constants
- expressions inside brackets have the right precedence
- px and rem conversion
- vw and vh resolution

Probably mostly working:
- Operator precedence, it possibly breaks down in some cases

Missing:
- Percentages: more complicated as it depends on element height and maybe some other stuff
- Variables, to keep down amount of changes and because finding the value is not very well handled elsewhere in the code
- Functions like min, max, clamp...