Improve non-native color pickers, make them default
For some reason, rendering pretty much the exact same picker takes much more time when it's the native picker, for Chromium.

The browsers needs multiple tens of milliseconds of scripting to output the widget, the same component done by the library in JS takes a few milliseconds.

Additionally, it requires less clicks to use, as it stays on the screen, and is very useful to visualize the current color.