WIP incrementally tracking references and reducer effects
This commit does not do anything, but it's nice to capture the initial point of the approach taken here.

The idea is to create an effect mechanism for reducers specifically.

This would fulfill the following need:

Currently the vast majority of time rendering, in many of the test content,
is spent figuring out references, very inefficiently.
It would be nice to index these references and update them incrementally, along with the editor state.

In case of a reducer, it's nice to request for an object that has each reducer key.
That way it's easy to create a "checklist" of how to handle each transition as to keep the derived state correct.

To really make this work with state that changes fast in the timeline,
it either needs to keep many snapshots to start from,
or the handler should become a pair of "apply/unapply" handlers.

Interestingly, I had to remove some TypeScript to more complete type results.