WIP improve full height view
Use CSS transforms to move fixed elements along with the fake window boundaries.

The same for sticky elements, but with some logic to start moving later and stop moving earlier.

A lot is not accounted for yet (margins, paddings, grid or flex containers, ...),
causing the result to be off, but usually not a lot.

Also provide an option to hide these elements.

Allow changing scale of this frame.