Fix incorrect calculation of :is and :not
I probably could have also updated the dependency...
In any case it should now get the correct result and the performance is ok.

The plugin seems to be using too much regex anyway and
it seems specificity could be calculated more efficiently. This can probably be done with the in progress parser.