Clip allowed ranges, fix incorrect chroma gradient
The picker should now produce (very close to) exactly the same result as the online picker.

A link to this picker is also added using url params.

The only thing "missing" is clipping the gaps in allowed hues for the other settings.

The previously used gradient for chroma was roughly correct,
but the lightness values didn't align along the way.
The gradient was made from gray to the fullest chroma color with no stops, which caused it to shift into other colors.
I guess it's because it's walking a straight line, where actually the color space has a detour between those colors. Or visce versa.

To fix this, I added color stops until I was sure the result was unnoticeable. This seems to not be too heavy to render.