More robust way to generate lightness picker gradient
The previous version (going from 0% to 100% lightness in all cases, relying on clipping the "wrong" parts) has the same effect on chromium based browsers, but has a flaw in Samsung browser, which has to be used instead of chromium if you want access to p3 colors...

That browser ended up displaying a grayscale, because probably gamut clipping is applied to the gradient colors at an earlier point before passing the color stops to whatever creates the gradient.

This is solved by changing 0% and 100% to the min and max ligthness for both the color stop's lightness and x-offset.