Interact with frame sync, Improve caching for inspections, avoid much double work
This is a pretty big refactor of how inspections are handled.

In the previous version, there were a lot of inefficiencies surrounding the inspections:

* The most expensive things in the init script ran twice for no good reason.
* All inspections were immediately hydrated for the entire history which made refreshing really slow.
* The entire result was stored in an array every time, duplicating all common groups.
* All of this was sent as a message which involved serializing and deserializing. This is simply not needed and can be done sync.

All these things are resolved in this commit by storing the groups using a WeakMap at various levels and removing the message passing.

The memory usage is now a lot more sane as storing duplicate groups
was the main source of wasted memory.

The only compromise is that the cached data will stick to the element much more than before. So reinspecting an accordion that was opened in the meanwhile will still consider it closed.

Element locators will be done in a later commit, its inspect button is now broken.