Ensure only "newer" nodes are immediately inspected when loading the page
Previously, loading the history list would cause all inspections to be executed at once, which is really slow. This happens because the visualization is trying to use the inspection data.

To fix this, the history visualization now only attempts to load data if  the element is cached or ahead of the current history position.

This allows to still use the visualization to get cache clearing out of the way in one interaction, by navigating to the initial state and opening the visualization. Doing the same by navigating history can take quite long and result in a very choppy experience.