Add "Xray" component that only shows the inspected element zoomed in
Show instance of page where all except the inspected tree is hidden.

`display: none` turned out more practical than `visibility: hidden` for this.

The latter is more accurate, but resulted too often in the element not properly framed. It's also a lot less efficient. With display, the only drawback is differences in CSS counters and exact text and image positions, which is fluid in the real world anyway.

Use outlines around direct and indirect descendants.
This was too visually disruptive to do in the main preview frame

Add a way to inspect next parent

Add some info and navigation